gestion de mensajeria

+++ src/app/components/admin/gestion-mensajeria/gestion-mensajeria.component.ts
@@ -1,7 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { QRCodeComponent } from 'angularx-qrcode'
-import { interval } from 'rxjs';
 @Component({
   selector: 'app-gestion-mensajeria',
   imports: [CommonModule, QRCodeComponent],
@@ -14,12 +13,10 @@ export class GestionMensajeriaComponent {
   qrLeido: boolean = false;
   intervalId: any;
   qrCode: string = '';
+  abortController: AbortController | null = null;
 
-  mostrarModal(mensaje: string): void {
-    alert(mensaje)
-  }
-  
-  async verificar(): Promise<void> {
+   
+  async verificar(mostrarMensajes: boolean): Promise<boolean> {
     try {
       const response = await fetch('http://localhost:3050/api/whatsapp/obtenerClientes', {
         method: 'GET',
@@ -35,24 +32,58 @@ export class GestionMensajeriaComponent {
 
       const data = await response.json();
       console.log(data);
-
+      
       if (data.clients && data.clients.length > 0) {
-        this.mensaje = 'Existe una sesion para los mensajes.';
-        this.estiloMensaje = 'success';
+        if (mostrarMensajes) {
+          this.mensaje = 'Existe una sesión para los mensajes.';
+          this.estiloMensaje = 'success';
+          setTimeout(() => {
+            this.mensaje = '';
+            this.estiloMensaje = '';
+          }, 10000);
+        }
+        return true;
       } else {
-        this.mensaje = 'No se encuentra ninguna sesion. Debes activarla lo antes posible';
-        this.estiloMensaje = 'warning';
+        if (mostrarMensajes) {
+          this.mensaje = 'No se encuentra ninguna sesión activa. Debes activar una lo antes posible';
+          this.estiloMensaje = 'warning';
+          setTimeout(() => {
+            this.mensaje = '';
+            this.estiloMensaje = '';
+          }, 10000);
+        }
+        return false;
       }
     } catch (error) {
       console.error('Error de sesion:', error);
-      this.mensaje = 'Error al obtener la sesion';
-      this.estiloMensaje = 'error';
+      if (mostrarMensajes) {
+        this.mensaje = 'Error al obtener la sesión';
+        this.estiloMensaje = 'error';
+        setTimeout(() => {
+          this.mensaje = '';
+          this.estiloMensaje = '';
+        }, 10000);
+      }
+      return false;
     }
   }
 
   
 
   async eliminar(): Promise<void> {
+    const sesionActiva = await this.verificar(false);
+
+    if (!sesionActiva) {
+      console.log('No hay sesion activa');
+      this.mensaje = 'Sesión actual eliminada. Recuerda crear una nueva lo antes posible.';
+      this.estiloMensaje = 'success';
+      setTimeout(() => {
+        this.mensaje = '';
+        this.estiloMensaje = '';
+      }, 10000);
+      return;
+    }
+
     await fetch('http://localhost:3050/api/whatsapp/eliminarClientes', {
       method: 'DELETE',
       headers: {
@@ -63,9 +94,12 @@ export class GestionMensajeriaComponent {
       .then(response => {
         if (response.ok) {
           console.log('Sesion eliminada correctamente');
-          this.mensaje = 'Sesion actual eliminada, recuerda crear una nueva lo antes posible';
+          this.mensaje = 'Sesión actual eliminada. Recuerda crear una nueva lo antes posible.';
           this.estiloMensaje = 'success';
-
+          setTimeout(() => {
+            this.mensaje = '';
+            this.estiloMensaje = '';
+          }, 10000);
         } else {
           throw new Error('No se pudo eliminar la sesión')
         }
@@ -73,47 +107,103 @@ export class GestionMensajeriaComponent {
       .then(data => console.log('Eliminar response:', data))
       .catch(error => {
         console.error('Error:', error);
-        this.mensaje = 'Ocurrio un error al eliminar la sesión, intentalo de nuevo mas tarde';
+        this.mensaje = 'Ocurrió un error al eliminar la sesión, intentalo de nuevo más tarde.';
         this.estiloMensaje = 'error';
+        setTimeout(() => {
+          this.mensaje = '';
+          this.estiloMensaje = '';
+        }, 10000);
       });
   }
 
-  actualizarQR(): void {
-    this.intervalId  = setInterval(() => {
-      if (this.qrLeido) {
-        this.qrCode = this.generarNuevoQR();
-      } else {
-        clearInterval(this.intervalId);
-      }
-    }, 5000 );
-  }
+  async nuevoSSE(): Promise<void> {
+    const sesionActiva = await this.verificar(false);
 
-  generarNuevoQR(): string {
-    return this.qrCode + Math.random().toString(36).substring(2, 5)
-  }
-
-  nuevo(): void {
+    if (sesionActiva) {
+      console.log('Sesion activa');
+      this.mensaje = 'Existe una sesión. Debes eliminarla para crear una nueva.';
+      this.estiloMensaje = 'warning';
+      setTimeout(() => {
+        this.mensaje = '';
+        this.estiloMensaje = '';
+      }, 10000);
+      return;
+    }
+    const body = JSON.stringify({
+      sessionId: "1234"
+    })
+    this.abortController = new AbortController();
     fetch('http://localhost:3050/api/whatsapp/CrearCliente', {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
         'Origin': '181.129.218.198'
       },
-      body: JSON.stringify({
-        sessionId: "1234"
-      }),
+      body: body,
       mode: 'cors',
-      credentials: 'omit'
-    })
-      .then(response => response.json())
-      .then(data => {
-        this.qrCode = data.qr;
-        this.actualizarQR();
-        console.log(this.qrCode)
-      })
-      .catch(error => {
-        console.error('Error:', error);
-        this.mensaje = 'Ocurrió un error al generar el QR.';
-      });
+      credentials: 'omit',
+      signal: this.abortController.signal
+    }).then(response => {
+      if (!response.body) {
+        throw new Error('El servidor no envio un cuerpo');
+      }
+      const reader = response.body.getReader();
+      const decoder = new TextDecoder('utf-8');
+
+      const readStream = () => {
+        reader.read().then(({ done, value}) => {
+          if (done) {
+            console.log('Conexion cerrada');
+            return;
+          }
+          const chunk = decoder.decode(value, { stream: true});
+          console.log('Datos recibidos:', chunk);
+
+          if (chunk.includes('Código QR generado:')) {
+            const qrCodeck = chunk.split('Código QR generado:')[1].trim();
+            this.qrCode = qrCodeck;
+            console.log('QR actualizado:', this.qrCode)
+          }
+
+          if (chunk.includes('Sesión 1234 lista')) {
+            console.log('Sesión detectada, cerrando conexion...');
+            this.mensaje = 'Sesión detectada exitosamente';
+            this.estiloMensaje = 'success';  
+            setTimeout(() => {
+              this.mensaje = '';
+              this.estiloMensaje = '';
+            }, 10000);
+            this.qrCode = '';
+            reader.cancel();
+            this.abortController = null;
+            return;
+          }
+
+          readStream();
+        }).catch(error => {
+          console.error('Error al leer el flujo:', error);
+        });
+      };
+      readStream();
+    }).catch(error => {
+      if (error.name === 'AbortError') {
+        console.log('Conexion cancelada por el cliente');
+      } else {
+        console.error('Error en la conexion', error);
+        this.mensaje = 'Ocurrió un error al conectar';
+      }
+    });
+  }
+
+  detener(): void {
+    if(this.abortController) {
+      this.abortController.abort();
+      this.abortController = null;
+      console.log('Conexion detenida')
+    }
+  }
+
+  ngOnDestroy(): void {
+    this.detener();
   }
 }
